requirements.txt
pandas
numpy
geopy
openpyxl
jupyter
geopy
folium
pypdf2
pdfplumber
camelot-py[cv]
scipy
seaborn